fixes #3: LocationConstraint on CreateBucketConfiguration
throw away S3Region, as S3 is not so entwined with regions
introduce enumeration to specify the location constraint when a bucket is created. US Standard uses an empty string as the location constraint.
The library uses the s3.amazonaws.com endpoint and doesn’t bother with region specific endpoints.