first draft of ConcurrentBatchWriter
An adaptive algorithm to write items in batch to DynamoDB, similar to the single-threaded batch writer.
A pool of writers using a matching pool of connections write concurrently to a single DynamoDB table
the writers share a variable that holds the number of milliseconds that each writer will pause for before attempting to write.
This pool of writers is fed batches of writes through a blocking buffer.
Collections of writes can be groups together. These groups are tagged with metadata, and one can check that all writes in a group have been run to completion.
Amazon client and service errors are caught and the failed batches are reported on a user supplied queue.