Improved implementation for FutureTransfer.listenFor
Cast to internal type `AbstractTransfer` so that a
low-level state change listener can be attached.
In light of issue #30, this provides a more foolproof
implementation of listening to transfer events.
There is now a three-way race to complete the
promise that signals the completion/termination of
the transfer.

This adds an integration test for s3, using the
fakes3 ruby gem.

closes #30